pyproject.toml
[tool.poetry]
name = "social-club-backend"
version = "0.1.0"
description = ""
authors = ["AbrarAlzubaidi <abraralzobaide98@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
django-allauth = "^0.47.0"
django-crispy-forms = "^1.13.0"
django-debug-toolbar = "^3.2.4"
whitenoise = "^5.3.0"
django-multiselectfield = "^0.1.12"
Pillow = "^8.4.0"
django-mysql = "^4.3.0"
docker = "^5.0.3"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.0.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.3"
django-environ = "^0.8.1"
django-cors-headers = "^3.10.1"
heroku = "^0.1.4"
dj-rest-auth = "^2.2.1"
fire = "^0.4.0"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
